Delete library service entirely
