+++ src-tauri/libcalibre/src/application/services/library/mod.rs
@@ -1,2 +1,1 @@
 pub mod dto;
-pub mod service;
